refactor(coordinator): split liveness tests into their own module
The added documentation and characterization tests pushed liveness.rs to
57052 bytes, over the 51200-byte Server Guards ceiling. Move the test
module into liveness_tests.rs behind #[path], the same convention
respawn_guard.rs and build_lease_reclaim.rs already use.

No test or production logic changed; 41 tests pass unchanged.
